requirements.txt
geopy==2.4.1
numpy==1.26.2
pandas==2.1.3
plotly==5.18.0
plotly-express==0.4.1
psycopg2-binary==2.9.9
py==1.11.0
py-dactyl==2.0.4
pyarrow==14.0.1
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dotenv==1.0.0
seaborn==0.13.0
SQLAlchemy==2.0.23
streamlit==1.29.0
streamlit-elements==0.1.0
uptime-kuma-api==1.2.1
